Convert guessgame into a proper package
Split nowplaying/guessgame.py (1817 lines) and
nowplaying/guessgamesettings.py (317 lines) into a package:

* guessgame/db.py — database management functions
* guessgame/scoring.py — pure text matching and scoring functions
* guessgame/server.py — server communication mixin
* guessgame/settings.py — settings UI (was guessgamesettings.py)
* guessgame/__init__.py — GuessGame class, slimmed to ~1100 lines

No behavior changes. All existing import paths preserved via
__init__.py delegates. Update callers in config.py and settingsui.py
to use nowplaying.guessgame.settings.
