+++ nowplaying/guessgame/settings.py
